java/src/main/java/br/com/cwi/api/security/service/BuscarUsuarioService.java
package br.com.cwi.api.security.service;

import br.com.cwi.api.security.domain.Usuario;
import br.com.cwi.api.security.repository.UsuarioRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

import static org.springframework.http.HttpStatus.NOT_FOUND;

@Service
public class BuscarUsuarioService {

    @Autowired
    private UsuarioRepository usuarioRepository;

    public Usuario porId(Long id){
        return usuarioRepository.findById(id)
                .orElseThrow( () -> new ResponseStatusException(
                        NOT_FOUND, "Usuário inválido."));
    }

}